requirements.txt
gunicorn
requests
python-dotenv
dash
dash-bootstrap-components
openai